# Make perf scripts ready for insecure scale-out tests