… cluster/gce/config-default.sh