@chenqianfzh 2021-01-29
> love seeing validation in the PR description. please also run the scale-up 100-node test to make sure of no regression.



> love seeing validation in the PR description. please also run the scale-up 100-node test to make sure of no regression.

Confirmed that scale-up scenario with a 100-node density run.